+++ packages/genesys/docs/stories/Hash/hashToShareData.tsx
@@ -1,0 +1,5 @@
+export default function hashToShareData(text: string): ShareData {
+    return {
+        text: `Check out this hash: ${text}`,
+    };
+}