+++ packages/genesys/packages/react-native-components/src/util/BlockchainAddress/BlockchainAddress.types.ts
@@ -1,13 +1,14 @@
 import { OverridableStringUnion } from "@peersyst/react-types";
-import { HashEllipsis } from "@peersyst/react-utils";
-import { TypographyProps } from "../../display/Typography";
 import {
     BlockchainLinksTypes,
     BlockchainLinksTypesOverrides,
 } from "@peersyst/react-components-core";
+import { HashProps } from "../Hash";
 
-export interface BlockchainAddressProps
-    extends Omit<TypographyProps, "numberOfLines" | "children"> {
+export type BlockchainAddressProps = Omit<
+    HashProps,
+    "numberOfLines" | "children" | "hash" | "hashToSharePayload"
+> & {
     /**
      * Blockchain address
      */
@@ -17,15 +18,7 @@ export interface BlockchainAddressProps
      */
     type: OverridableStringUnion<BlockchainLinksTypes, BlockchainLinksTypesOverrides>;
     /**
-     * Address' length in characters
+     * Address to sharePayload mapper
      */
-    length?: number;
-    /**
-     * If ellipsis should be in the middle or end
-     */
-    ellipsis?: HashEllipsis;
-    /**
-     * Show Icon copy
-     */
-    copy?: boolean;
-}
+    addressToSharePayload?: HashProps["hashToSharePayload"];
+};