+++ packages/genesys/packages/react-components/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 3.6.0 **BREAKING**
+
+[Hash] Refactor hash to work with actions: `share`, `copy`, `link`, `customAction`
+[BlockchainAddress] Refactor to work with new Hash component
+
 ## 3.5.4
 
 [useNumericInput] Fix useNumericInput